chore: re-export relocated components
Components relocated to their own modules were
re-exported.'
