--- src/components/select/type.ts
@@ -1,24 +1,0 @@
-/* eslint-disable unused-imports/no-unused-vars */
-export type SelectOption = {
-  label: string;
-  value: string | number;
-};
-
-type MultipleSelectProps = {
-  multiple: true;
-  value: SelectOption[];
-  onChange: (value: SelectOption[]) => void;
-};
-
-type SingleSelectProps = {
-  multiple?: false;
-  value?: SelectOption;
-  onChange: (value: SelectOption | undefined) => void;
-};
-
-export type SelectProps = {
-  options: SelectOption[];
-  classname?: string;
-  label?: string;
-  id: string;
-} & (SingleSelectProps | MultipleSelectProps);
