+++ src/components/select/types.ts
@@ -1,0 +1,8 @@
+import { SelectHTMLAttributes } from "react";
+import { SelectOption } from "features/listings/types";
+
+export interface NewSelectProps
+  extends SelectHTMLAttributes<HTMLSelectElement> {
+  options: SelectOption[];
+  label?: string;
+}
